fix: Nested NetworkBehaviours don't de-register or Invoke OnNetworkDespawn if destroyed while the parent NetworkObject remains spawned [NCCBUG-137] (#2091)
* fix

NCCBUG-137

This resolves the issue where destroying a GameObject with one or more NetworkBehaviour components that is a child or any child generation of the NetworkObject/GameObject that theNetworkBehaviour(s) are assigned to would still try to access the destroyed NetworkBehaviour since it was still an entry within the NetworkObject.ChildNetworkBehaviours list.  It also resolves the issue where under this scenario the NetworkBehaviours would not have their OnNetworkDespawn method invoked.

* update

NCCBUG-137

updating the changelog

* fix

We need to check IsSpawned when logging a warning about a NetworkBehaviour not having an assigned NetworkObject (this is only valid if NetworkBehaviour.IsSpawned == true)

* fix

Needed to adjust to account for the condition when it is valid is when you have a NetworkBehaviour that is spawned but there is no assigned NetworkObject.

* test

NCCBUG-137
This is the integration validation test for this PR fix.

* update

Adding PR number to the changelog entry

* test

adding check for OnNetworkDespawn being invoked when a child NetworkBehaviour is destroyed. but the parent NetworkObject persists and remains spawned.

* style

MTT-4260
was created for this technical debt.
removing reminder to create a ticket.

* fix

fixing merge issue.

* fix

Removing the invocation of NetworkBehaviour's OnNetworkDespawn when a NetworkBehaviour is destroyed while a NetworkObject is still spawned.

* style

removing unused namespace.

* Update com.unity.netcode.gameobjects/CHANGELOG.md

Co-authored-by: Jesse Olmer <jesseo@unity3d.com>

Co-authored-by: Jesse Olmer <jesseo@unity3d.com>