fix: player objects do not honor the DontDestroyWithOwner property [MTT-4458] (#2225)
* fix
This fixes the issue with player objects not honoring the DontDestroyWithOwner property.
The summary for DontDestroyWithOwner was incorrect.  When true the NetworkObject's ownership is given back to the server.

* test
Adding a test to verify that when a client disconnects and the client's player object has DontDestroyWithOwner set to true that the NetworkObject's ownership is transferred back to the client.