fix: NetworkAnimator rising RTT issue (#2236)
* fix
This fixes the issue where NetworkAnimator was not completely updated to the changes with AnimationMessage.  NetworkAnimator will now only send states that have changed.
Removed the check for AnimationMessage.AnimationStates being null as it will always be null when deserializing.
Updated several comments for readability and clarity.

* test
adding additional checks to assure or AnimationStates list does not increase in size over time.