feat: Supporting managed types in NetworkVariable [MTT-4594] (#2219)
* feat: Supporting managed types in NetworkVariable.

Also changes NetworkVariable to deserialize INetworkSerializable types in-place - this is done to avoid having a GC alloc every time the value changes. There will only be a GC alloc if it receives an updated value and the current value is null.

Managed types are supported in two ways:

1. By making them INetworkSerializable, or
2. By assigning the delegates UserNetworkVariableSerialization<T>.WriteValue and UserNetworkVariableSerialization<T>.ReadValue

If a managed type is passed in without meeting one of those criteria (or any unsupported type), then an exception will be thrown the first time the NetworkVariable is read or written. That timing is to ensure no weird behavior with exceptions being thrown before user code is able to assign the delegates.

Co-authored-by: Fatih Mar <mfatihmar@gmail.com>